// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  STUDENT: 'STUDENT',
  CLUB_MEMBER: 'CLUB_MEMBER',
  CLUB_ADMIN: 'CLUB_ADMIN',
  CLUB_OWNER: 'CLUB_OWNER',
  COLLEGE_ADMIN: 'COLLEGE_ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ClubCategory = {
  TECH: 'TECH',
  CULTURAL: 'CULTURAL',
  SPORTS: 'SPORTS',
  SOCIAL: 'SOCIAL',
  ACADEMIC: 'ACADEMIC',
  ARTS: 'ARTS',
  OTHER: 'OTHER'
} as const

export type ClubCategory = (typeof ClubCategory)[keyof typeof ClubCategory]


export const ClubStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  SUSPENDED: 'SUSPENDED',
  REJECTED: 'REJECTED'
} as const

export type ClubStatus = (typeof ClubStatus)[keyof typeof ClubStatus]


export const ClubMemberRole = {
  FOLLOWER: 'FOLLOWER',
  MEMBER: 'MEMBER',
  ADMIN: 'ADMIN',
  OWNER: 'OWNER'
} as const

export type ClubMemberRole = (typeof ClubMemberRole)[keyof typeof ClubMemberRole]


export const EventType = {
  IN_PERSON: 'IN_PERSON',
  ONLINE: 'ONLINE',
  HYBRID: 'HYBRID'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]


export const EventStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const RegistrationStatus = {
  CONFIRMED: 'CONFIRMED',
  WAITLISTED: 'WAITLISTED',
  CANCELLED: 'CANCELLED',
  ATTENDED: 'ATTENDED'
} as const

export type RegistrationStatus = (typeof RegistrationStatus)[keyof typeof RegistrationStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const NotificationType = {
  EVENT_REGISTRATION: 'EVENT_REGISTRATION',
  EVENT_REMINDER: 'EVENT_REMINDER',
  EVENT_UPDATE: 'EVENT_UPDATE',
  EVENT_CANCELLED: 'EVENT_CANCELLED',
  CLUB_ANNOUNCEMENT: 'CLUB_ANNOUNCEMENT',
  CLUB_INVITE: 'CLUB_INVITE',
  CHECK_IN_OPEN: 'CHECK_IN_OPEN',
  CERTIFICATE_READY: 'CERTIFICATE_READY',
  WAITLIST_PROMOTED: 'WAITLIST_PROMOTED',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]